Support for `web` projects
If the downloaded client secret is for a `web` application instead of an `installed` application, the object in the JSON file will be called `web` instead of `installed`, which causes `auth.js` to crash (see https://developers.google.com/api-client-library/python/guide/aaa_client_secrets for more information). This commit adds support for both types of client secrets. Tested and works well.